+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Features
 - Add `neptune_project_member` resource ([#2](https://github.com/AleksanderWWW/terraform-provider-neptune/pull/2))
+- Added updating project member role ([#8](https://github.com/AleksanderWWW/terraform-provider-neptune/pull/8))
+
 
 ### Changes
 - If resources in Neptune.ai already exist (when creating) or do not exist (when deleting) - only warn ([#3](https://github.com/AleksanderWWW/terraform-provider-neptune/pull/3))